Create film card buttons component
